+++ tests/integration.rs
@@ -247,7 +247,7 @@ fn test_app() -> Result<(), Box<dyn Error>> {
     std::fs::create_dir(&stubs_dir).unwrap();
 
     let bundled_libs_search_arg =
-        "-L".to_owned() + current_dir()?.join("tapHLE_dylibs").to_str().unwrap();
+        "-L".to_owned() + current_dir()?.join("runtime/dylibs").to_str().unwrap();
     let stubs_lib_search_arg = "-L".to_owned() + stubs_lib_dir.to_str().unwrap();
     let stubs_frameworks_search_arg = "-F".to_owned() + stubs_frameworks_dir.to_str().unwrap();
     let mut extra_linker_args = Vec::<String>::new();